requirements.txt
# Core dependencies
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
scipy>=1.11.0

# Vision backbones
timm>=1.0.3
transformers>=4.40.0
open_clip_torch>=2.24.0

# Simulation (install separately: pip install mani-skill==3.0.0b22 sapien==3.0.0b1 --pre)
# mani-skill>=3.0.0b22
# sapien>=3.0.0b1

# Dynamics similarity
tslearn>=0.6.3
dtaidistance>=2.3.10

# Diffusion Policy
diffusers>=0.27.0
einops>=0.7.0

# Evaluation
# robomimic>=0.3.0  # Optional: pip install robomimic (needs cmake + EGL)
gymnasium>=0.29.0

# Visualization & analysis
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
umap-learn>=0.5.5
plotly>=5.18.0

# Logging & tracking
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Image metrics
lpips>=0.1.4
pytorch-fid>=0.3.0

# Config & utilities
hydra-core>=1.3.2
omegaconf>=2.3.0
pyyaml>=6.0
rich>=13.7.0

# Video metrics
opencv-python>=4.9.0

# Distributed training
accelerate>=0.27.0
# deepspeed>=0.14.0  # Optional: pip install deepspeed (needs CUDA dev headers)

# Data handling
h5py>=3.10.0
lmdb>=1.4.1
webdataset>=0.2.86
